requirements.txt
# Core - Web Scraping
selenium>=4.15.0
webdriver-manager>=4.0.0

# Data Processing
pandas>=2.2.0
yfinance>=0.2.36

# Configuration & Validation
python-dotenv>=1.0.0
pydantic>=2.5.0

# CLI & Logging
rich>=13.7.0
loguru>=0.7.0

# Testing (optional)
pytest>=7.4.0
